+++ frontend/src/pages/ReportPage.jsx
@@ -1,23 +1,33 @@
 // contexts
 import useBugsContext from "../hooks/useBugsContext";
+import useAuthContext from "../hooks/useAuthContext";
 // react router dom elements 
-import { useNavigate, useLoaderData, useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 // date-fns to format date correctly
 import { formatDistanceToNow } from "date-fns";
 // components
 import Comment from "../components/Comment";
+import { useEffect, useState } from "react";
 
 export default function ReportPage() {
-    const report = useLoaderData();
+    const [report, setReport] = useState(null);
     const { reportId } = useParams();
     const { dispatch } = useBugsContext();
+    const { user } = useAuthContext();
     // navigate to a path
     const navigate = useNavigate();
 
+    // this handles the deletion on the report page
     const handleDelete = async () => {
+        // exit out of the function if the user is not authenticated
+        if (!user) {
+            return;
+        }
+
         // send a post request to delete the report
         const response = await fetch(`http://localhost:3000/api/bugs/${reportId}`, {
             method: 'DELETE',
+            headers: { 'Authorization': `Bearer ${user.token}` }
         });
         const json = await response.json();
 
@@ -28,19 +38,41 @@ export default function ReportPage() {
 
     };
 
+    useEffect(() => {
+        const getReportDetails = async () => {
+            // fetch the data inside a try catch block
+            const response = await fetch(`http://localhost:3000/api/bugs/${reportId}`, {
+                headers: { "Authorization": `Bearer ${user.token}` },
+            });
+            const json = await response.json();
+
+            // else throw an error
+            if (!response.ok) throw Error(json.error);
+            // if response is ok return the data
+            setReport(json);
+        }
+
+        if (user) {
+            getReportDetails();
+        }
+    }, [user]);
+
     return (
         <div className="report-page">
-            <div className="container" >
-                <div className="info">
-                    <p>Username asked:</p>
-                    <span onClick={handleDelete} className="material-symbols-outlined">
-                        delete
-                    </span>
+            {!report && <div>loading...</div>}
+            {
+                report && (<div className="container" >
+                    <div className="info">
+                        <p>Username asked:</p>
+                        <span onClick={handleDelete} className="material-symbols-outlined">
+                            delete
+                        </span>
+                    </div>
+                    <p className="date">{formatDistanceToNow(new Date(report.createdAt), { addSuffix: true })}</p>
+                    <h2>{report.title}</h2>
+                    <p className="description">{report.description}</p>
                 </div>
-                <p className="date">{formatDistanceToNow(new Date(report.createdAt), { addSuffix: true })}</p>
-                <h2>{report.title}</h2>
-                <p className="description">{report.description}</p>
-            </div>
+                )}
 
             <h2 className="section-title">Comments:</h2>
             {/* comments section */}
@@ -56,14 +88,15 @@ export default function ReportPage() {
     )
 }
 
-export async function reportDetailsLoader({ params }) {
-    const { reportId } = params;
-    // fetch the data inside a try catch block
-    const response = await fetch(`http://localhost:3000/api/bugs/${reportId}`);
-    const json = await response.json();
+// loader function to get the specific report details
+// export async function reportDetailsLoader({ params }) {
+//     const { reportId } = params;
+//     // fetch the data inside a try catch block
+//     const response = await fetch(`http://localhost:3000/api/bugs/${reportId}`);
+//     const json = await response.json();
 
-    // else throw an error
-    if (!response.ok) throw Error(json.error);
-    // if response is ok return the data
-    return json;
-}
+//     // else throw an error
+//     if (!response.ok) throw Error(json.error);
+//     // if response is ok return the data
+//     return json;
+// }
